Struct vizia_winit::application::Application

pub struct Application { /* private fields */ }

Creating a new application creates a root Window and a Context. Views declared within the closure passed to Application::new() are added to the context and rendered into the root window.

Example

Application::new(|cx|{
   // Content goes here
})
.run();

Calling run() on the Application causes the program to enter the event loop and for the main window to display.

Implementations

impl Application

pub fn new<F>(content: F) -> Self
where F: 'static + FnOnce(&mut Context),

pub fn ignore_default_theme(self) -> Self

Sets the default built-in theming to be ignored.

pub fn should_poll(self) -> Self

pub fn on_idle<F: 'static + Fn(&mut Context)>(self, callback: F) -> Self

Takes a closure which will be called at the end of every loop of the application.

The callback provides a place to run ‘idle’ processing and happens at the end of each loop but before drawing. If the callback pushes events into the queue in state then the event loop will re-run. Care must be taken not to push events into the queue every time the callback runs unless this is intended.

Example

Application::new(|cx| {
    // Build application here
})
.on_idle(|cx| {
    // Code here runs at the end of every event loop after OS and vizia events have been handled
})
.run();

pub fn get_proxy(&self) -> ContextProxy

Returns a ContextProxy which can be used to send events from another thread.

pub fn run(self) -> Result<(), ApplicationError>