Module vizia_winit::rwh

Structs

Active

Keeps track of whether the application is currently active.

ActiveHandle

Represents a live window handle.

AndroidDisplayHandle

Raw display handle for Android.

AndroidNdkWindowHandle

Raw window handle for Android NDK.

AppKitDisplayHandle

Raw display handle for AppKit.

AppKitWindowHandle

Raw window handle for AppKit.

DisplayHandle

The handle to the display controller of the windowing system.

DrmDisplayHandle

Raw display handle for the Linux Kernel Mode Set/Direct Rendering Manager.

DrmWindowHandle

Raw window handle for the Linux Kernel Mode Set/Direct Rendering Manager.

GbmDisplayHandle

Raw display handle for the Linux Generic Buffer Manager.

GbmWindowHandle

Raw window handle for the Linux Generic Buffer Manager.

HaikuDisplayHandle

Raw display handle for Haiku.

HaikuWindowHandle

Raw window handle for Haiku.

OrbitalDisplayHandle

Raw display handle for the Redox operating system.

OrbitalWindowHandle

Raw window handle for the Redox operating system.

UiKitDisplayHandle

Raw display handle for UIKit.

UiKitWindowHandle

Raw window handle for UIKit.

WaylandDisplayHandle

Raw display handle for Wayland.

WaylandWindowHandle

Raw window handle for Wayland.

WebDisplayHandle

Raw display handle for the Web.

WebWindowHandle

Raw window handle for the Web.

Win32WindowHandle

Raw window handle for Win32.

WinRtWindowHandle

Raw window handle for WinRT.

WindowHandle

The handle to a window.

WindowsDisplayHandle

Raw display handle for Windows.

XcbDisplayHandle

Raw display handle for Xcb.

XcbWindowHandle

Raw window handle for Xcb.

XlibDisplayHandle

Raw display handle for Xlib.

XlibWindowHandle

Raw window handle for Xlib.

Enums

HandleError

The error type returned when a handle cannot be obtained.

RawDisplayHandle

A display server handle for a particular windowing system.

RawWindowHandle

A window handle for a particular windowing system.

Traits

HasDisplayHandle

A display that acts as a wrapper around a display handle.

HasRawDisplayHandle

Display that wraps around a raw display handle.

HasRawWindowHandle

Window that wraps around a raw window handle.

HasWindowHandle

A handle to a window.